% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sex_inference.R
\name{cache_sex_signals}
\alias{cache_sex_signals}
\title{Cache pre-computed sex signals (called from input_processing.R when
SigDFs are processed in streaming fashion so we don't have to retain
SigDFs in memory).}
\usage{
cache_sex_signals(sex_signal_df, scale = c("intensity", "beta"))
}
\description{
Cache pre-computed sex signals (called from input_processing.R when
SigDFs are processed in streaming fashion so we don't have to retain
SigDFs in memory).
}
\keyword{internal}
