% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clock_computation.R
\name{check_clock_availability}
\alias{check_clock_availability}
\title{Check which clock packages are available}
\usage{
check_clock_availability(available_probes, verbose = TRUE)
}
\description{
Check which clock packages are available
}
\keyword{internal}
