% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clock_computation.R
\name{compute_all_clocks}
\alias{compute_all_clocks}
\title{Compute all clocks}
\usage{
compute_all_clocks(betas, pheno = NULL, n_cores, verbose = TRUE)
}
\arguments{
\item{betas}{Beta matrix (CpGs as rows, samples as columns)}

\item{pheno}{Optional phenotype data.frame}

\item{n_cores}{Number of CPU cores}

\item{verbose}{Print progress}
}
\value{
Data frame with all clock results
}
\description{
Compute all clocks
}
\keyword{internal}
