% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{detect_array_platform}
\alias{detect_array_platform}
\title{Detect array platform from probe IDs}
\usage{
detect_array_platform(probe_ids)
}
\description{
Detect array platform from probe IDs
}
\keyword{internal}
