% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{determine_optimal_cores}
\alias{determine_optimal_cores}
\title{Determine optimal number of cores}
\usage{
determine_optimal_cores(n_samples, n_probes, requested_cores, verbose)
}
\description{
The "* 3" multiplier on per-sample memory accounts for: (1) input matrix,
(2) intermediate copy used during clock projection, (3) scratch space for
matrix-multiply temporaries.
}
\keyword{internal}
