% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clock_coefficients.R
\name{.bocklandt_inline}
\alias{.bocklandt_inline}
\title{Bocklandt 3-CpG saliva clock (Bocklandt et al. 2011, PLoS One 6:e14821)
Linear regression: age = sum_i b_i * beta_i + intercept}
\usage{
.bocklandt_inline()
}
\description{
Bocklandt 3-CpG saliva clock (Bocklandt et al. 2011, PLoS One 6:e14821)
Linear regression: age = sum_i b_i * beta_i + intercept
}
\keyword{internal}
