% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clock_coefficients.R
\name{.garagnani_inline}
\alias{.garagnani_inline}
\title{Garagnani 1-CpG ELOVL2 clock (Garagnani et al. 2012, Aging Cell 11:1132)}
\usage{
.garagnani_inline()
}
\description{
Garagnani 1-CpG ELOVL2 clock (Garagnani et al. 2012, Aging Cell 11:1132)
}
\keyword{internal}
