% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clock_coefficients.R
\name{.vidal_bralo_inline}
\alias{.vidal_bralo_inline}
\title{Vidal-Bralo 8-CpG clock (Vidal-Bralo et al. 2016, Front Genet 7:126)}
\usage{
.vidal_bralo_inline()
}
\description{
Vidal-Bralo 8-CpG clock (Vidal-Bralo et al. 2016, Front Genet 7:126)
}
\keyword{internal}
