% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.R
\name{get_probe_chromosomes}
\alias{get_probe_chromosomes}
\title{Get chromosome information for probes from manifest}
\usage{
get_probe_chromosomes(platform, verbose = TRUE)
}
\description{
Get chromosome information for probes from manifest
}
\keyword{internal}
