% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_processing.R
\name{load_input_data}
\alias{load_input_data}
\title{Load beta values from various input formats}
\usage{
load_input_data(input, n_cores, verbose)
}
\description{
Load beta values from various input formats
}
\keyword{internal}
