% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{log_msg}
\alias{log_msg}
\title{Log message with formatting}
\usage{
log_msg(fmt, ..., verbose = TRUE)
}
\arguments{
\item{fmt}{Format string (passed to sprintf)}

\item{...}{Arguments for format string}

\item{verbose}{Logical. Print message if TRUE.}
}
\description{
Log message with formatting
}
\keyword{internal}
