% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sex_inference.R
\name{ortho_dist}
\alias{ortho_dist}
\title{Orthogonal distance from points to a fitted line}
\usage{
ortho_dist(x, y, fit)
}
\description{
Orthogonal distance from points to a fitted line
}
\keyword{internal}
