% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{qc_save}
\alias{qc_save}
\title{Save an R object using qs2 if available, fall back to RDS}
\usage{
qc_save(object, path)
}
\description{
Save an R object using qs2 if available, fall back to RDS
}
\keyword{internal}
