% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_betas}
\alias{validate_betas}
\title{Validate beta matrix format and values}
\usage{
validate_betas(betas)
}
\arguments{
\item{betas}{Matrix to validate}
}
\value{
Validated (possibly corrected) beta matrix
}
\description{
Robust M-value detection: only flips M->beta when the \emph{median} (not the
min/max) of the input falls outside the unit interval (0 to 1). A
single bad probe in 800k shouldn't wreck an entire matrix.
}
\keyword{internal}
